+++ tests/unit/docs_snippets/test_docs_snippets.py
@@ -1,8 +1,8 @@
 import re
 
 
-def test_guide_snippets(cpp_test):
-    output = cpp_test("guide_snippets_test.cpp")
+def test_docs_snippets(cpp_test):
+    output = cpp_test("docs_snippets_test.cpp")
     match = re.search(r"TEST done (\d+)/(\d+)", output)
     assert match, output
     assert match.group(1) == match.group(2), output